feat(gotjunk): Move navigation buttons to floating left sidebar
## Summary
- Created new LeftSidebarNav component with floating buttons on left side
- Removed right section navigation buttons from BottomNavBar
- BottomNavBar now only contains camera controls and search bar
- All 4 tab buttons (Browse, Map, My Items, Cart) + Liberty Alert moved to left sidebar

## Changes
1. **LeftSidebarNav.tsx** (new): Vertical floating navigation on left side
   - Browse (Grid icon) - Tab 1 with SOS detection
   - Map (Map icon) - Tab 2
   - My Items (Home icon) - Tab 3
   - Cart (Cart icon) - Tab 4
   - Liberty Alert (🗽) when unlocked

2. **BottomNavBar.tsx**: Simplified to camera controls only
   - Removed right section (lines 253-320)
   - Removed unused imports (GridIcon, MapIcon, HomeIcon, CartIcon)
   - Kept search bar and camera orb with arrows

3. **App.tsx**: Integrated both components
   - Added LeftSidebarNav import
   - Split props between LeftSidebarNav and BottomNavBar
   - Moved all tab navigation logic to left sidebar

## UI/UX Impact
- Cleaner bottom navigation focused on camera
- Easier thumb access to navigation on left side
- Consistent floating button design with backdrop blur
- Active tab highlighting preserved (blue glow ring)
